Fix checkout action from account cart pop out
This fixes a client-reported bug in the checkout process from the
account portion of the site. The checkout function attempts to access
an Upscribe function that doesn't exist in the account page context,
causing an error and preventing the user from checking out.

The fix is to add the Upscribe cart logic to the account template,
which makes the function available and allows the customer to get to
checkout.