Add Development theme setting for dev server URL
Adds the `dev_url` Theme setting to provide the URL for the local
development server. This fixes issues with injecting local script
changes (running under Theme Kit and Browsersync) with the template
changes running from Shopify (uploaded by Theme Kit). Previously,
this was relying on the entire site being served from the local domain
but that setup causes issues with account login and cart behavior as
the cookies that are used are scoped to the store domain and unavailable
from `localhost` or `lvh.me` or whatever we use.