Add update payment button to SMI page
Adds the update payment button to the SMI page. Typically, this button,
and the associated code, is added via the Ordergroove subscription page.
Since we're embedding this page in the application UI, we need to add
the button and the submission handling manually.

This adds the button and the code to the js module to handle the form
submission using the authentication information provided in the liquid
template.