Fix product-form getPriceForOptionValue function
Fixes a bug in the `getPriceForOptionValue` function where it would
reset to the "selected" variant and use that price for all options
in all buttons. For example, if the 1lb was $19.99, then all the button
prices would be reset to $19.99 after a product was added to the cart.

Since this function is only called on the product form and only called
for the price deciding option, we can make better assumptions about
how to use the data to find the variant prices. This option is the
variant option used as the `priceDecidingFactor` - in other words, it's
the factor (option) used to decide the product price displayed.

This updates the function to use a simple lookup on the variant using
the passed in option and removes the previous logic that considered
the selected option.