Update og-offer to handle missing selling plan map
Client reported a bug where a new error reporting tool was detecting
some issues accessing an `undefined` object on product listing pages.
The pages with the error were products that didn't have a subscription
plan available for the default selected variant. The code that attempts
to update the pricing on the buttons was not checking for the plan
before accessing it.

This updates the `og-offer` module to check for a selling plan before
attempting to update the button prices.