WIP - refactor how og-offer looks up the ref keys for the selling plans
This updates the functions to use a little indirection rather than
building the refkey directly. The idea is that we can now change how
the keys are built and what they contain without hard-coding. Should
make it easier to change the behavior to support a few different
versions of lookups.